+++ packages/block-editor/src/private-apis.js
@@ -56,7 +56,6 @@ import {
 	mediaSideloadFromUrlKey,
 	openMediaEditorModalKey,
 } from './store/private-keys';
-import { requiresWrapperOnCopy } from './components/writing-flow/utils';
 import { PrivateRichText } from './components/rich-text/';
 import { PrivateBlockPopover } from './components/block-popover';
 import { PrivateInserterLibrary } from './components/inserter/library';
@@ -126,7 +125,6 @@ lock( privateApis, {
 	globalStylesDataKey,
 	globalStylesLinksDataKey,
 	selectBlockPatternsKey,
-	requiresWrapperOnCopy,
 	PrivateRichText,
 	PrivateInserterLibrary,
 	reusableBlocksSelectKey,