+++ packages/block-library/src/list-item/index.js
@@ -1,6 +1,5 @@
 import { __ } from '@wordpress/i18n';
 import { listItem as icon } from '@wordpress/icons';
-import { privateApis } from '@wordpress/block-editor';
 import { privateApis as blocksPrivateApis } from '@wordpress/blocks';
 import initBlock from '../utils/init-block';
 import metadata from './block.json';
@@ -26,7 +25,6 @@ export const settings = {
 		};
 	},
 	transforms,
-	[ unlock( privateApis ).requiresWrapperOnCopy ]: true,
 	__experimentalLabel( attributes, { context } ) {
 		const { content } = attributes;